refactor: move link control
