+++ packages/controls/src/index.ts
@@ -5,6 +5,7 @@ export * from './color'
 export * from './combobox'
 export * from './icon-radio-group'
 export * from './image'
+export * from './link'
 export * from './number'
 export * from './text-input'
 export * from './text-area'
